Real animation - note: if you run this in emulator, you may not see the day-dusk-night color changes
The first step will be to animate the sun down below the horizon. Inside onCreateView(...), pull out a couple of views into fields on SunsetFragment.

Write code to animate the sun. Here is the plan: smoothly move mSunView so that its top is right at the edge of the top of the sea. The way that you will do this is to translate the location of the top of mSunView to the bottom of its parent.

The first step is to find where the animation should start and end. Write this first step in a new method called startAnimation().

The getTop() method is one of four methods on View that return the local layout rect for that view: getTop(), getBottom(), getRight(), and getLeft(). A view's local layout rect is the position and size for that view in relation to its parent, as was determined when that view was laid out. It is possible to change the location of the view on screen by modifying these values, but it is not recommended. They are reset every time a layout pass occurs, so
they tend not to hold their value.
In any event, the animation will start with the top of the view at its current location. It needs to end with the top at the bottom of mSunView's parent, mSkyView. To get it there, it should be as far down as mSkyView is tall, which you can find by calling getHeight(). The getHeight() method returns the same thing as getTop() - getBottom().

Now that you know where the animation should start and end, create and run an ObjectAnimator to perform it. Then hook up startAnimation() --- in onCreateView() --- so that it is called every time the user taps anywhere in the scene